# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сокращения URL",
        "title": "Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/user/urls": {
            "get": {
                "summary": "Просмотр коротких ссылок пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseView"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "summary": "Удаление ссылок пользователем",
                "parameters": [
                    {
                        "description": "объект с сылками для удаления",
                        "name": "Delete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ResponseView": {
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string"
                },
                "short_Url": {
                    "type": "string"
                }
            }
        },
...
    }
}